tor-hsservice: Make num_dequeued u32 instead of usize.
We need to divide a Duration by this, and Duration only supports
division by u32.

This is, in a sense, just pushing around where the overflow would
happen, (from the conversion in the suggested update to the increment
when a item is dequeued). However, this overflow is so unlikely to
happen (it would require more that 14 million requests per second) that
it does not seem worth slowing down the increment operation to try to
handle it.

+++ crates/tor-hsservice/src/pow/v1.rs
@@ -714,7 +714,7 @@ struct RendRequestReceiverInner<R, Q> {
     /// greater than or equal to the suggested effort.
     num_enqueued_gte_suggested: usize,
     /// Number of requests that were dequeued during the current update period.
-    num_dequeued: usize,
+    num_dequeued: u32,
     /// Amount of time during the current update period that we spent with no requests in the
     /// queue.
     idle_time: Duration,
@@ -772,7 +772,7 @@ impl<R: Runtime, Q: MockableRendRequest + Send + 'static> RendRequestReceiver<R,
 
         if inner.num_dequeued != 0 {
             let update_period_duration = inner.runtime.now() - inner.update_period_start;
-            let avg_request_duration = update_period_duration / inner.num_dequeued as u32;
+            let avg_request_duration = update_period_duration / inner.num_dequeued;
             if inner.queue.is_empty() {
                 let now = inner.runtime.now();
                 let last_transition = inner.last_transition;
